+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [0.12.6-dev.1] - 2018-02-13
+
 - `bit status` shows a list of staged versions in 'staged components' section 
 - introduced a new command `bit untag` for reverting un-exported tags. 
 - init local scope inside .git